# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 WeChatFerry RPC 实现的微信接口，使用 Go 语言编写，无第三方运行时依赖，易于对接任意编程语言。",
        "title": "Wrest Chat Api",
        "contact": {
            "name": "WeChatRest",
            "url": "https://github.com/opentdp/wrest-chat"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v0.10.0"
    },
    "basePath": "/",
    "paths": {
...
        "/wcf/accept_new_friend": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::联系人管理"
                ],
                "summary": "接受好友请求",
                "parameters": [
                    {
                        "description": "接受好友参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.AcceptNewFriendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wcfrest.CommonPayload"
                        }
                    }
                }
            }
        },
        "/wcf/add_chatroom_members": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::群聊管理"
                ],
                "summary": "添加群成员",
                "parameters": [
                    {
                        "description": "管理群成员参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.ChatroomMembersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wcfrest.CommonPayload"
                        }
                    }
                }
            }
        },
        "/wcf/alias_in_chatroom": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::群聊管理"
                ],
                "summary": "获取群成员昵称",
                "parameters": [
                    {
                        "description": "获取群成员昵称参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.GetAliasInChatRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/wcf/audio_msg": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::消息收取"
                ],
                "summary": "获取语音消息",
                "parameters": [
                    {
                        "description": "获取语音消息参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.GetAudioMsgRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wcfrest.CommonPayload"
                        }
                    }
                }
            }
        },
        "/wcf/avatars": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::联系人管理"
                ],
                "summary": "获取头像列表",
                "parameters": [
                    {
                        "description": "获取头像列表参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.GetAvatarsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/wcfrest.AvatarPayload"
                            }
                        }
                    }
                }
            }
        },
        "/wcf/chatroom_members": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::群聊管理"
                ],
                "summary": "获取群成员列表",
                "parameters": [
                    {
                        "description": "获取群成员列表参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.GetChatRoomMembersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/wcfrest.ContactPayload"
                            }
                        }
                    }
                }
            }
        },
...
        "/wcf/db_query_sql": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::数据库查询"
                ],
                "summary": "执行数据库查询",
                "parameters": [
                    {
                        "description": "数据库查询参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.DbSqlQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/wcf/db_tables": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::数据库查询"
                ],
                "summary": "获取数据库表列表",
                "parameters": [
                    {
                        "description": "获取数据库表列表参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.GetDbTablesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/wcfrest.DbTablePayload"
                            }
                        }
                    }
                }
            }
        },
        "/wcf/del_chatroom_members": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::群聊管理"
                ],
                "summary": "删除群成员",
                "parameters": [
                    {
                        "description": "管理群成员参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.ChatroomMembersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wcfrest.CommonPayload"
                        }
                    }
                }
            }
        },
...
        "/wcf/download_attach": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::消息收取"
                ],
                "summary": "下载附件",
                "parameters": [
                    {
                        "description": "下载附件参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.DownloadAttachRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wcfrest.CommonPayload"
                        }
                    }
                }
            }
        },
        "/wcf/download_image": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::消息收取"
                ],
                "summary": "下载图片",
                "parameters": [
                    {
                        "description": "下载图片参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.DownloadImageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wcfrest.CommonPayload"
                        }
                    }
                }
            }
        },
...
        "/wcf/forward_msg": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::消息发送"
                ],
                "summary": "转发消息",
                "parameters": [
                    {
                        "description": "转发消息参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.ForwardMsgRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wcfrest.CommonPayload"
                        }
                    }
                }
            }
        },
...
        "/wcf/invite_chatroom_members": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::群聊管理"
                ],
                "summary": "邀请群成员",
                "parameters": [
                    {
                        "description": "管理群成员参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.ChatroomMembersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wcfrest.CommonPayload"
                        }
                    }
                }
            }
        },
...
        "/wcf/ocr_result": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::消息收取"
                ],
                "summary": "获取OCR识别结果",
                "parameters": [
                    {
                        "description": "获取OCR识别结果参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.GetOcrRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wcfrest.CommonPayload"
                        }
                    }
                }
            }
        },
        "/wcf/receive_transfer": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WCF::消息收取"
                ],
                "summary": "接受转账",
                "parameters": [
                    {
                        "description": "接受转账参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wcfrest.ReceiveTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wcfrest.CommonPayload"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "wcfrest.AcceptNewFriendRequest": {
            "type": "object",
            "properties": {
                "scene": {
                    "description": "添加方式：17 名片，30 扫码",
                    "type": "integer"
                },
                "v3": {
                    "description": "加密的用户名",
                    "type": "string"
                },
                "v4": {
                    "description": "验证信息 Ticket",
                    "type": "string"
                }
            }
        },
...
        "wcfrest.ChatroomMembersRequest": {
            "type": "object",
            "properties": {
                "roomid": {
                    "description": "群聊 id",
                    "type": "string"
                },
                "wxids": {
                    "description": "用户 id 列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "wcfrest.CommonPayload": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误信息"
                },
                "result": {
                    "description": "返回结果",
                    "type": "string"
                },
                "success": {
                    "description": "是否成功",
                    "type": "boolean"
                }
            }
        },
...
        "wcfrest.GetAliasInChatRoomRequest": {
            "type": "object",
            "properties": {
                "roomid": {
                    "description": "群聊 id",
                    "type": "string"
                },
                "wxid": {
                    "description": "用户 id",
                    "type": "string"
                }
            }
        },
...
        "wcfrest.GetAvatarsRequest": {
            "type": "object",
            "properties": {
                "wxids": {
                    "description": "用户 id 列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}